Api connected to supabase

+++ api/src/main/java/com/example/api/Repository/CustomerRepository.java
@@ -1,5 +1,8 @@
 package com.example.api.Repository;
-import com.example.api.Model.Customer;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
+import com.example.api.Model.Customer;
+
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
